% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondaryscripts.R
\name{DownloadCSV}
\alias{DownloadCSV}
\title{Download Traffic Count Data to .CSV File}
\usage{
DownloadCSV(data)
}
\arguments{
\item{data}{The data that the user wishes to aggregate.}
}
\value{
A CSV file for each site is downloaded, with the naming convention of
"X_DDMMMYYYY-DDMMMYYYY.csv", where "X" denotes the alphanumeric characters of the site's
geographic address, "DD" denotes the start and end date days respectively, "MMM" denotes the month
as three letters for the start and end date months respectively, and "YYYY" denotes the year
for the start and end date years respectively.
}
\description{
Motorway Incident Detection and Automatic Signalling (MIDAS) is a
sensor based network along UK motorways, and is designed to collect data regarding
traffic flows, average speeds and road occupancy, amongst other features, on the
road network. This data can be accessed via the MIDAS website
\url{https://www.midas-data.org.uk/} (login required) in the form of .tcd.bz2 files.

The function \code{DownloadCSV} is designed to download data found using the
\link[ScrapeMIDAS]{RoadData} into an appropriate .CSV format.
}
\examples{
DownloadCSV(trafficData)

}
\author{
RAC Foundation, Luke Hudlass-Galley
}
