% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondaryscripts.R
\name{GeneratePercentiles}
\alias{GeneratePercentiles}
\title{Estimate Speed Percentiles for Traffic Count Data}
\usage{
GeneratePercentiles(data, percentile, timeInterval)
}
\arguments{
\item{data}{The data that the user wishes to investigate.}

\item{percentile}{Numeric; the percentile of drivers that travel under a particular speed
(a value commonly investigated 85\%; that is, the function will find out the speed at which 85%
of drivers travel at or below).}

\item{timeInterval}{Numeric; the MIDAS data does not provide accurate details regarding the speed
of every vehicle travelling, only the average speed and the number of vehicles passing a MIDAS site
per minute. The function therefore assumes that all vehicles in a minute travel at the same average
speed, and so to calculate a percentile speed, a sample must be selected over a period of time. This
argument is that period of time, in minutes. This time interval must divide the number of rows in the
\code{data} dataframe.}
}
\value{
A version of the format of \code{data}, where additional data has been added in the form of
a column at the end of the data, denoting the percentile speed. Please note that all rows that are in
the same time interval share the same value in this column. For example, if 60 minute samples are taken,
then the first 60 rows will share the same value, followed by the second 60 rows, and so on.
}
\description{
Motorway Incident Detection and Automatic Signalling (MIDAS) is a
sensor based network along UK motorways, and is designed to collect data regarding
traffic flows, average speeds and road occupancy, amongst other features, on the
road network. This data can be accessed via the MIDAS website
\url{https://www.midas-data.org.uk/} (login required) in the form of .tcd.bz2 files.

The function \code{GeneratePercentiles} provides the user with an estimated \eqn{n}th percentile speed,
(the speed that \eqn{n}\% of the drivers within a given time period drive at or below) based on the data
collected in the \link[ScrapeMIDAS]{RoadData}.

Please note that this function only provides an estimate of the percentile speeds, due to the
nature of the MIDAS data. It should be acknowledged that the results from this function may not
be representative of the true percentile speeds.
}
\examples{
dataWithHourly85Percentiles <- GeneratePercentiles(roadData, 85, 60)

}
\author{
RAC Foundation, Luke Hudlass-Galley
}
