% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondaryscripts.R
\name{PCUAddition}
\alias{PCUAddition}
\title{Incorporation Passenger Car Units to Traffic Count Data}
\usage{
PCUAddition(data, category1Unit, category2Unit, category3Unit, category4Unit)
}
\arguments{
\item{data}{The data that the user wishes to aggregate.}

\item{category1}{Numeric; the user-selected PCU value for category 1 of vehicles in MIDAS data.}

\item{category2}{Numeric; the user-selected PCU value for category 2 of vehicles in MIDAS data.}

\item{category3}{Numeric; the user-selected PCU value for category 3 of vehicles in MIDAS data.}

\item{category4}{Numeric; the user-selected PCU value for category 4 of vehicles in MIDAS data.}
}
\value{
A version of the format of \code{data}, where additional data has been added in the form of
a column at the end of the data, denoting the PCU value for each row of \code{data}.
}
\description{
Motorway Incident Detection and Automatic Signalling (MIDAS) is a
sensor based network along UK motorways, and is designed to collect data regarding
traffic flows, average speeds and road occupancy, amongst other features, on the
road network. This data can be accessed via the MIDAS website
\url{https://www.midas-data.org.uk/} (login required) in the form of .tcd.bz2 files.

The function \code{PCUAddition} allows the user to add a "passenger car unit" column
to the data collected in the \link[ScrapeMIDAS]{RoadData}, with user defined values for
each of the four vehicle categories.
}
\examples{
JanuaryDataWithPCU <- PCUAddition(JanuaryData, 1, 1.5, 2, 2.5)

}
\author{
RAC Foundation, Luke Hudlass-Galley
}
