% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondaryscripts.R
\name{PlotData}
\alias{PlotData}
\title{Plotting Tools for Traffic Count Data}
\usage{
PlotData(data, timeRange, n, attributesXName, ...)
}
\arguments{
\item{data}{The data that the user wishes to examine.}

\item{timeRange}{The start and end minute of the data, formatted as "start minute:end minute".
These are measured from the start of the dataset. For example, "1:1440" denotes the whole of the first
day, whereas "5761:7200" denotes the whole of the fifth day.}

\item{n}{The plotting 'character' that is plotted on the graph (the same as "pch" in \link[graphics]{points}).}

\item{attributesXName}{The (quoted) name of the column found in the data that the user wishes to plot on the x-axis.
Note that to denote "Time" on the axis, use "" (2 quotation marks) as the value for this argument.}

\item{...}{Any number of  (quoted) names of columns found in the data that the user wishes to plot on the y-axis.}
}
\value{
A plot of the requested data.
}
\description{
Motorway Incident Detection and Automatic Signalling (MIDAS) is a
sensor based network along UK motorways, and is designed to collect data regarding
traffic flows, average speeds and road occupancy, amongst other features, on the
road network. This data can be accessed via the MIDAS website
\url{https://www.midas-data.org.uk/} (login required) in the form of .tcd.bz2 files.

The function \code{PlotData} is designed to make plotting data found using the
\link[ScrapeMIDAS]{RoadData} more accessible.
}
\examples{
PlotData(trafficData, 1:1440, ".", "", "Average Speed (Lane 1)")
PlotData(roadData2016, 1380:1440, 20, "Average Speed (Lane 2)", "Occupancy (Lane 3)")

}
\author{
RAC Foundation, Luke Hudlass-Galley
}
