% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondaryscripts.R
\name{SelectRushHour}
\alias{SelectRushHour}
\title{Isolate or Remove Rush Hour Times with Traffic Count Data}
\usage{
SelectRushHour(data, isolate)
}
\arguments{
\item{data}{The data that the user wishes to isolate or remove the rush hour data from.}

\item{isolate}{A paramater, which if equal to TRUE, will return just the rush hour data.
If set to FALSE, it will return the data that is not during rush hour.}
}
\value{
A version of the format of \code{data}, either with just the rush hour data or
just the data that is not during rush hour, depending on if the \code{isolate} parameter
is set to TRUE or FALSE respectively.
}
\description{
Motorway Incident Detection and Automatic Signalling (MIDAS) is a
sensor based network along UK motorways, and is designed to collect data regarding
traffic flows, average speeds and road occupancy, amongst other features, on the
road network. This data can be accessed via the MIDAS website
\url{https://www.midas-data.org.uk/} (login required) in the form of .tcd.bz2 files.

The function \code{RushHourFunction} allows the user to either remove or isolate the
rush hour periods, defined from 07:30 - 09:30 and 16:00 - 18:00 each day.
}
\examples{
M25DataRushHour <- SelectRushHour(M25Data, isolate = TRUE)

}
\author{
RAC Foundation, Luke Hudlass-Galley
}
